% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_lexrank.R
\name{bind_lexrank_}
\alias{bind_lexrank}
\alias{bind_lexrank_}
\title{Bind lexrank scores to a dataframe of sentences}
\usage{
bind_lexrank_(tbl, sentence, doc_id, sent_id, threshold = 0.2,
  usePageRank = TRUE, damping = 0.85, continuous = FALSE, ...)

bind_lexrank(tbl, sentence, doc_id, sent_id, threshold = 0.2,
  usePageRank = TRUE, damping = 0.85, continuous = FALSE, ...)
}
\arguments{
\item{tbl}{dataframe containing column of sentences to be lexranked}

\item{sentence}{name of column containing sentences to be lexranked}

\item{doc_id}{name of column containing document ids corresponding to the sentences column}

\item{sent_id}{name of column containing sentence ids corresponding to the sentences column}

\item{threshold}{The minimum simil value a sentence pair must have to be represented in the graph where lexRank is calculated.}

\item{usePageRank}{\code{TRUE} or \code{FALSE} indicating whether or not to use the page rank algorithm for ranking sentences.  If \code{FALSE}, a sentences unweighted centrality will be used as the rank.  Defaults to \code{TRUE}.}

\item{damping}{The damping factor to be passed to page rank algorithm.  Ignored if \code{usePageRank} is \code{FALSE}.}

\item{continuous}{\code{TRUE} or \code{FALSE} indicating whether or not to use continuous LexRank.  Only applies if \code{usePageRank==TRUE}.  If \code{TRUE}, \code{threshold} will be ignored and lexRank will be computed using a weighted graph representation of the sentences. Defaults to \code{FALSE}.}

\item{...}{tokenizing options to be passed to lexRankr::tokenize}
}
\value{
A dataframe with an additional column of lexrank scores (column is given name lexrank)
}
\description{
Bind lexrank scores to a dataframe of sentences
}
\examples{
library(dplyr)

df <- dplyr::tibble(doc_id = 1:3, 
                    text = c("Testing the system. Second sentence for you.", 
                             "System testing the tidy documents df.", 
                             "Documents will be parsed and tagged."))

df \%>\% 
  unnest_sentences(sents, text) \%>\% 
  bind_lexrank(sents, doc_id, sent_id)
  
df \%>\% 
  unnest_sentences(sents, text) \%>\% 
  bind_lexrank_("sents", "doc_id", "sent_id")
}

