% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence_parser.R
\name{sentence_parser}
\alias{sentence_parser}
\title{Utility to parse sentences from text}
\usage{
sentence_parser(text)
}
\arguments{
\item{text}{Character vector to be parsed into sentences}
}
\value{
A list with length equal to `length(text)`; list elements are character vectors of text parsed with sentence regex
}
\description{
Utility to parse sentences from text; created to have a central shared sentence parsing function
}
