% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smart_stopwords}
\alias{smart_stopwords}
\title{SMART English Stopwords}
\format{a character vector with 571 elements}
\source{
\url{http://jmlr.csail.mit.edu/papers/volume5/lewis04a/}
}
\usage{
smart_stopwords
}
\description{
English stopwords from the SMART information retrieval system (as documented in Appendix 11 of \url{http://jmlr.csail.mit.edu/papers/volume5/lewis04a/})
}
\keyword{datasets}
